Move keyboard interop to KeyboardHelper
